pyproject.toml
[project]
name = "parentpass-chatbot-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
dependencies = [
    "aiohttp>=3.9.0",
    "baml-py==0.202.0",
    "fastapi>=0.115.12",
    "google-cloud-bigquery>=3.34.0",
    "pydantic>=2.11.4",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.27.0",
    "coverage>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "--verbose"

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/baml_client/*"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.venv
  | baml_client
  | analytics_reports
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = [
    "E203",  # whitespace before ':'
    "E501",  # line too long (handled by black)
    "W503",  # line break before binary operator
    "W504",  # line break after binary operator
]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "baml_client",
    "analytics_reports",
    "*.egg-info",
]
per-file-ignores = [
    "__init__.py:F401",  # unused imports in __init__.py files
]
